requirement.txt
# requirements.txt

Pillow==9.0.1
pyheif==0.5.0
openpyxl==3.0.9
tkinter  # tkinter通常随Python安装，无需单独安装